Complete Category-service

+++ category-service/src/main/java/com/evan/categoryservice/service/impl/CategoryServiceImpl.java
@@ -2,6 +2,7 @@ package com.evan.categoryservice.service.impl;
 
 import com.evan.categoryservice.dto.CategoryDTO;
 import com.evan.categoryservice.entity.Category;
+import com.evan.categoryservice.exception.ResourceNotFoundException;
 import com.evan.categoryservice.mapper.CategoryMapper;
 import com.evan.categoryservice.repository.CategoryRepository;
 import com.evan.categoryservice.service.CategoryService;
@@ -27,15 +28,17 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public CategoryDTO getCategoryById(Long categoryId) {
         Category category = categoryRepository.findById(categoryId).get();
+        if (category == null) throw new ResourceNotFoundException("Category", "id", categoryId);
         CategoryDTO categoryDTO = CategoryMapper.mapToCategoryDto(category);
         return categoryDTO;
     }
 
     @Override
-    public List<CategoryDTO> getCategoriesByName(String name) {
-        List<Category> categories = categoryRepository.findCategoryByName(name);
-        List<CategoryDTO> categoriesDTO = categories.stream().map(CategoryMapper::mapToCategoryDto).collect(Collectors.toList());
-        return categoriesDTO;
+    public CategoryDTO getCategoryByName(String name) {
+        Category category = categoryRepository.findCategoryByName(name);
+        if (category == null) throw new ResourceNotFoundException("Category", "name", name);
+        CategoryDTO categoryDTO = CategoryMapper.mapToCategoryDto(category);
+        return categoryDTO;
     }
 
     @Override
@@ -47,7 +50,14 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public CategoryDTO updateCategory(CategoryDTO categoryDTO) {
-        Category category = CategoryMapper.mapToCategory(categoryDTO);
+        Category category = categoryRepository.findCategoryByName(categoryDTO.getName());
+        if (category == null) throw new ResourceNotFoundException("Category", "name", categoryDTO.getName());
+        Category convertToCategory = CategoryMapper.mapToCategory(categoryDTO);
+
+        category.setName(convertToCategory.getName());
+        category.setDescription(convertToCategory.getDescription());
+        category.setPostIds(convertToCategory.getPostIds());
+
         Category newCategory = categoryRepository.save(category);
         CategoryDTO newCategoryDto = CategoryMapper.mapToCategoryDto(newCategory);
         return newCategoryDto;
@@ -56,6 +66,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public String deleteCategoryById(Long categoryId) {
         Category delCategory = categoryRepository.findById(categoryId).get();
+        if (delCategory == null) throw new ResourceNotFoundException("Category", "id", categoryId);
         categoryRepository.delete(delCategory);
         return "delete success";
     }
